style: reorganize package structures
